Enable AD dynamic DNS updates
Added variable documentation to README.md, variables with comments
to defaults/main.yml, and single blockinfile task to update
/etc/sssd/sssd.conf with the needed options to enable and configure
dynamic updating of A, AAAA, and PTR records in AD's DNS service.

Included are all documented options for tuning how the updates
happen with the exception of dyndns_auth_ptr, and
dyndns_update_per_family.

dyndns_auth_ptr was excluded because it isn't supported by SSSD
in all of the versions of RHEL that this role covers. When not
configured it defaults to the same value as dyndns_auth.

dyndns_update_per_family while mentioned in the man pages of
sssd-ad(5), was rejected on every version of RHEL tested.

Signed-off-by: Jeremy Winder <jwinder@redhat.com>