fix: use command stdin for password, and do not log password
Cause: The code was constructing the realm join command to be passed
via the shell module, including piping the password into the command,
and was showing the command, including the password, when using
check mode.

Consequence: The clear text password was available in the logs when
using check mode.

Fix: Use command with stdin for the password instead of shell.  The
password is not part of the command.  command with stdin is more
secure than using shell.  The debug output has been changed to
show the command with the `ad_integration_join_parameters` removed,
because we cannot know if those parameters contain data which should
not be logged.  Those parameters will still be passed to the actual
realm join command.

Result: The password is not logged.  The role is more secure.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>