refactor: use the ini_file module to test sssd.conf
The `ini` lookup is problematic with ansible 2.9 due to the
arguments being completely different, and because ansible attempts
to evaluate `{{ lookup('ini', ...) }}` even if guarded by a `when`,
any use of the new style 2.x `ini` arguments will fail.
Instead, just use the `ini_file` module itself, and rely on the fact
that it will report `changed: true` if the values are different.

Use `community.general.ini_file` instead of `ini_file` everywhere
for consistency.

Refactor the `tests_dyndns.yml` test so that it works in CI with
a single host - in that case, just check that the sssd.conf was
written correctly.

Rename tests_sssd_custom_setttings.yml to tests_sssd_custom_settings.yml

Use the new `ad_integration_sssd_custom_settings` in tests instead of
writing to sssd.conf directly with `ini_file`

The `tests_dyndns.yml` test wasn't working properly - in order for
any parameters to be set, you must specify `ad_dyndns_update: true`

Use non-FQCN `win_command` in a couple of places that Ansible 2.9 was
giving an error about.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>