feat: search for name of domain/realm in sssd.conf; merge settings if duplicates
Feature: Search for the name of the section used in the SSSD config file
for the domain/realm specific settings, as managed by `ad_dyndns_update` and
`ad_integration_sssd_custom_settings`.  This section is added by `realm join`
and the casing of the name is determined by AD settings, so we need to search
for the name using a case insensitive search.
Previous versions of the role would always use a lower case realm in the section
name, which may have left the SSSD config file with multiple sections
for the realm.  Use the new `ad_integration_sssd_merge_duplicate_sections` setting
to consolidate all of the settings from the multiple sections into the chosen
section.

Reason: The realm/domain section in the SSSD config file is written by
`realm join`, and the case is determined by AD settings, so we must do a search
to find the section name to use for writting additional SSSD realm/domain
settings.  In addition, we need some way to clean up any previously
duplicated settings from the SSSD config.

Result: The ad_integration role can manage domain/realm sections in the SSSD
config file correctly.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>